+++ ipc/network_ipc.c
@@ -129,7 +129,7 @@ int ipc_get_status(ipc_message *msg)
 	return ret;
 }
 
-int ipc_inst_start_ext(sourcetype source, int len, char *buf)
+int ipc_inst_start_ext(sourcetype source, size_t len, char *buf)
 {
 	int connfd;
 	ipc_message msg;
